Moved temporary files from recharts dir to charts
