+++ src/visits/charts/DoughnutChartLegend.scss
