+++ src/visits/charts/DoughnutChartLegend.tsx
